Test for Arabic content, not match
Testing for matching wasn't catching errors. Adjacent lines in the dictionary
can have very similar content. A match test would require much more precision.
A simple solution to this problem is to test for Arabic content, since adjacent
lines do not usually both contain Arabic content, the presence of Arabic is
enough to determine whether the line is the one that should be replaced.